GUI development on AVR

+++ sharedLibs/DLComms/DLComms.cpp
@@ -56,17 +56,12 @@ void DLComms::Process()
 	while (this->Seriala.available())
 	{
 		uint8_t dataByte = this->Seriala.read();
-		// Serial.print(dataByte, HEX);
-		// Serial.print(" ");
-		if ((dataByte == static_cast<uint8_t>(DL_COMMAND::DATA) ||
-			 dataByte == static_cast<uint8_t>(DL_COMMAND::CURRENT_SET)) && 
+		if ((dataByte == DL_COMMAND::DATA ||
+			 dataByte == DL_COMMAND::CURRENT_SET) && 
 			 this->rxActive == false)
 		{
-			// this->DataRXbuffer[0] = dataByte;
 			this->rxActive = true;
-			// this->rxBufferIndex = 1;
 			this->rxStartMillis = millis();
-			// Serial.print("1");
 		}
 
 		if (this->rxActive == true)
@@ -75,17 +70,14 @@ void DLComms::Process()
 			{
 				this->rxActive = false;
 				this->rxBufferIndex = 0;
-				// Serial.print("2");
 				break;
 			}
 
-			// Serial.print("3");
 			this->DataRXbuffer[this->rxBufferIndex] = dataByte;
 			this->rxBufferIndex += 1;
 			if (this->rxBufferIndex >= sizeof(DataRXbuffer))
 			{
 				this->rxActive = false;
-				// Serial.print("4");
 				break;
 			}
 		}
@@ -95,30 +87,17 @@ void DLComms::Process()
 
 	if (this->rxBufferIndex == sizeof(DataRXbuffer))
 	{
-		// Serial.print("5");
 		auto *const currentPacket = reinterpret_cast<DLPacket_t *const>(this->DataRXbuffer);
 
 		uint32_t rxCRC = currentPacket->crc;
-		// Serial.println(rxCRC);
 
 		currentPacket->crc = 0;	// remove CRC from received packet to match source packet
 		uint32_t calculatedCRC = CRC32::calculate(this->DataRXbuffer, sizeof(DataRXbuffer));
 
-		// for (uint8_t n = 0; n < sizeof(DataRXbuffer); n++)
-		// {
-		// 	Serial.print(DataRXbuffer[n], HEX);
-		// 	Serial.print(" ");
-		// }
-		// Serial.println();
-
-		// Serial.println(calculatedCRC);
-
 		if (calculatedCRC == rxCRC)
 		{
-			// Serial.print("6");
 			if (currentPacket->ackRequested)
 			{
-				// Serial.print("7");
 				sendAck(currentPacket->packetNumber);	
 			}
 
@@ -127,37 +106,14 @@ void DLComms::Process()
 
 			if (currentPacket->Command == DL_COMMAND::DATA)
 			{
-				// Serial.print("8");
 				auto *const currentPacketData = reinterpret_cast<DLData_t *const>(currentPacket->dataBuffer);
 				if (this->DataReceiveFunction != NULL)
 				(*DataReceiveFunction)(currentPacketData);
 			}
-
-			// Serial.print("milliAmpsSetVal: ");
-			// Serial.println(valuePacket->currentSet);
-
-			// Serial.print("voltsReadVal: ");
-			// Serial.println(valuePacket->voltageRead);
-
-			// Serial.print("milliAmpsReadVal: ");
-			// Serial.println(valuePacket->currentRead);
-
-			// Serial.print("lastRX: ");
-			// Serial.println(lastRXmillis);
-
-			// Serial.print("millis: ");
-			// Serial.println(millis());
-
-			// Serial.print("diff: ");
-			// Serial.println(millis() - lastRXmillis);
-
-			// Serial.println();
 		}
 		else
 		{
-			// Serial.print("9");
 			this->crcErrorCounter++;
-			// Serial.println("#### CRC error ####");
 		}
 
 		this->rxBufferIndex = 0;
